+++ src/gameObjects/getLocalBounds.ts
@@ -3,6 +3,7 @@ import { IChild, ISizeable } from "../types";
 
 export function getLocalBounds(obj: ISizeable & IChild): Phaser.Geom.Rectangle {
   const worldBounds = obj.getBounds();
+  console.log('worldBounds', worldBounds);
   const localOffset = getLocalOffset(obj);
   return new Phaser.Geom.Rectangle(worldBounds.x - localOffset.x, worldBounds.y - localOffset.y, worldBounds.width, worldBounds.height);
 }